Reformatted Appjs file and created Uploader component
